feat: Solution for 00-basic-app
For this one everything was provided all I had to do was to make sure
I called `bootstrapApplication` in `main.ts`.
For this exercise everything was provided for the project structure.

All I had to do was to call `bootstrapApplication` in `main.ts`.

Calling this function starts the Angular engine so to speak. It takes
the root component of the application and renderes it. The root
component of the application has to be a standalone one.